third_party/WebKit/LayoutTests/svg/animations/svglength-animation-values-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test SVGLength animation set with 'values', different units and different count of spaces.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 130
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 151.2
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 576
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 267
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
